refactor: extract out aggregate state builder
