+++ lib/commanded/aggregates/aggregate.ex
@@ -100,17 +100,15 @@ defmodule Commanded.Aggregates.Aggregate do
   require Logger
 
   alias Commanded.Aggregates.Aggregate
+  alias Commanded.Aggregates.AggregateStateBuilder
   alias Commanded.Aggregates.ExecutionContext
   alias Commanded.Event.Mapper
   alias Commanded.Event.Upcast
   alias Commanded.EventStore
   alias Commanded.EventStore.RecordedEvent
-  alias Commanded.EventStore.SnapshotData
   alias Commanded.Registration
   alias Commanded.Snapshotting
   alias Commanded.Telemetry
-
-  @read_event_batch_size 100
 
   defstruct [
     :application,
@@ -231,7 +229,7 @@ defmodule Commanded.Aggregates.Aggregate do
   def handle_continue(:populate_aggregate_state, %Aggregate{} = state) do
     # Subscribe to aggregate's events to catch any events appended to its stream
     # by another process, such as directly appended to the event store.
-    {:noreply, populate_aggregate_state(state), {:continue, :subscribe_to_events}}
+    {:noreply, AggregateStateBuilder.populate(state), {:continue, :subscribe_to_events}}
   end
 
   @doc false
@@ -410,69 +408,6 @@ defmodule Commanded.Aggregates.Aggregate do
     end
   end
 
-  # Populate the aggregate's state from a snapshot, if present, and it's events.
-  #
-  # Attempt to fetch a snapshot for the aggregate to use as its initial state.
-  # If the snapshot exists, fetch any subsequent events to rebuild its state.
-  # Otherwise start with the aggregate struct and stream all existing events for
-  # the aggregate from the event store to rebuild its state from those events.
-  defp populate_aggregate_state(%Aggregate{} = state) do
-    %Aggregate{aggregate_module: aggregate_module, snapshotting: snapshotting} = state
-
-    aggregate =
-      case Snapshotting.read_snapshot(snapshotting) do
-        {:ok, %SnapshotData{source_version: source_version, data: data}} ->
-          %Aggregate{
-            state
-            | aggregate_version: source_version,
-              aggregate_state: data
-          }
-
-        {:error, _error} ->
-          # No snapshot present, or exists but for outdated state, so use intial empty state
-          %Aggregate{state | aggregate_version: 0, aggregate_state: struct(aggregate_module)}
-      end
-
-    rebuild_from_events(aggregate)
-  end
-
-  # Load events from the event store, in batches, to rebuild the aggregate state
-  defp rebuild_from_events(%Aggregate{} = state) do
-    %Aggregate{
-      application: application,
-      aggregate_uuid: aggregate_uuid,
-      aggregate_version: aggregate_version
-    } = state
-
-    case EventStore.stream_forward(
-           application,
-           aggregate_uuid,
-           aggregate_version + 1,
-           @read_event_batch_size
-         ) do
-      {:error, :stream_not_found} ->
-        # aggregate does not exist, return initial state
-        state
-
-      event_stream ->
-        rebuild_from_event_stream(event_stream, state)
-    end
-  end
-
-  # Rebuild aggregate state from a `Stream` of its events.
-  defp rebuild_from_event_stream(event_stream, %Aggregate{} = state) do
-    Enum.reduce(event_stream, state, fn event, state ->
-      %RecordedEvent{data: data, stream_version: stream_version} = event
-      %Aggregate{aggregate_module: aggregate_module, aggregate_state: aggregate_state} = state
-
-      %Aggregate{
-        state
-        | aggregate_version: stream_version,
-          aggregate_state: aggregate_module.apply(aggregate_state, data)
-      }
-    end)
-  end
-
   defp aggregate_lifespan_timeout(lifespan, function_name, args) do
     # Take the last event or the command or error
     args = args |> List.wrap() |> Enum.take(-1)
@@ -580,7 +515,7 @@ defmodule Commanded.Aggregates.Aggregate do
     else
       {:error, :wrong_expected_version} ->
         # Fetch missing events from event store
-        state = rebuild_from_events(state)
+        state = AggregateStateBuilder.rebuild_from_events(state)
 
         # Retry command if there are any attempts left
         case ExecutionContext.retry(context) do
